solving buf in parsing, add a Debug.cpp file
