+++ srcs/parser/test.conf
@@ -12,7 +12,7 @@ server
 	allow GET;
 
 	location /random_folder {
-		allow	GET DELETE POST;
+		allow		GET DELETE POST;
 		autoindex on;
 
 		root 	html;
@@ -24,7 +24,7 @@ server {
 	listen 127.0.0.1:443 default_server;
 
 	location / {
-		allow GET GHOST DELETE;
+		allow			GET GHOST DELETE;
 		autoindex off;
 
 		root 	html;
